Move fuzzing tests from encoding to a dedicated file
